requirements.txt
requests==2.7.0
nose==1.3.7
freezegun==0.3.5
pep8
python-dotenv